refactor(backend): move path-access helpers to src/helpers
Path validation utilities are general helpers, not agent-core internals.
