+++ apps/backend/src/agent-core/tool/index.ts
@@ -1,10 +1,4 @@
 export {loadSkillTool} from './load-skill.js';
-export {
-  AccessCheckResult,
-  checkAccess,
-  isSubPath,
-  isSubPathOrSelf,
-} from './path-access.js';
 export {ToolRegistry} from './tool-registry.js';
 export type {
   ShellState,
@@ -13,4 +7,10 @@ export type {
   ToolExecuteStatus,
   ToolExecutionContext,
 } from './types.js';
+export {
+  AccessCheckResult,
+  checkAccess,
+  isSubPath,
+  isSubPathOrSelf,
+} from '@/helpers/path-access.js';
 export type {AllowedPathEntry} from '@omnicraft/settings-schema';
