+++ apps/backend/src/helpers/path-access.ts
